fix(sdcm/remote): Fix remoter code duplication
